Moved contestants related components in components/contentants/
