+++ src/components/contestants/contestants.less
@@ -1,4 +1,4 @@
-@import "../main";
+@import "../../main";
 
 
 .contestants {
@@ -130,7 +130,7 @@
 
           .map-marker {
             float: left;
-            background: url("../../assets/img/icons/location.png");
+            background: url("../../../assets/img/icons/location.png");
             width: 20px;
             height: 20px;
             background-size: 20px 20px;
@@ -144,7 +144,7 @@
 
           .comments {
             float: right;
-            background: url("../../assets/img/icons/comments.png");
+            background: url("../../../assets/img/icons/comments.png");
             text-align: center;
             width: 25px;
             height: 25px;
@@ -218,7 +218,7 @@
 
           .comments {
             float: right;
-            background: url("../../assets/img/icons/comments.png");
+            background: url("../../../assets/img/icons/comments.png");
             text-align: center;
             width: 25px;
             height: 25px;
@@ -229,7 +229,7 @@
 
           .map-marker {
             float: left;
-            background: url("../../assets/img/icons/location.png");
+            background: url("../../../assets/img/icons/location.png");
             width: 20px;
             height: 20px;
             background-size: 20px 20px;
